pyproject.toml
[build-system]
requires = ["hatchling", "hatch-vcs"]
build-backend = "hatchling.build"

[project]
dynamic = ["version"]
name = "armaio"
authors = [
    {name="MrClock", email="mrclock8163@gmail.com"},
]
description = "IO modules for handling Arma 3 file formats."
readme = "README.md"
requires-python = ">=3.11"
dependencies = []
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: GNU General Public License v3 (GPLv3)",
    "Operating System :: OS Independent",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3.14",
    "Topic :: Games/Entertainment",
    "Topic :: Software Development",
    "Topic :: Software Development :: Libraries",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Typing :: Typed"
]
license = "GPL-3.0"
license-files = ["LICEN[CS]E*"]
keywords = ["arma3", "bi", "p3d", "rtm", "paa"]

[project.urls]
Repository = "https://github.com/MrClock8163/armaio"
Issues = "https://github.com/MrClock8163/armaio/issues"

[project.optional-dependencies]
all = ["armaio[pillow]"]
pillow = [
    "pillow~=12.0"
]

[dependency-groups]
testing = [
    "pytest ~= 9.0.0"
]
linting = [
    {include-group = "testing"},
    "mypy >= 1.17.1",
    "flake8 >= 7.3.0"
]
development = [
    {include-group = "linting"}
]
documentation = [
    "sphinx ~= 9.0.0",
    "sphinx-notfound-page ~= 1.1.0",
    "sphinx-last-updated-by-git ~= 0.3.8",
    "furo == 2025.12.19"
]

[tool.hatch.version]
source = "vcs"
fallback-version = "0.0.0"

[tool.hatch.version.raw-options]
local_scheme = "no-local-version"

[tool.hatch.build.hooks.vcs]
version-file = "src/armaio/_version.py"

[tool.pytest.ini_options]
addopts = ["-ra", "-v", "--color=yes"]
testpaths = ["tests"]

[tool.mypy]
strict = true
